Added unit tests and an option to set the processor grid

+++ cbenchio/src/dataGenerator.cpp
@@ -7,7 +7,7 @@ void indexDataGenerator::generate( distributedCartesianArray & disData)
     auto & localData = disData.getData();
     auto localShape = disData.getLocalShape();
     auto localOffset = disData.getLocalOffset();
-
+    
     for( index_t i =0 ; i< localShape[0] ; i++  )
     {
         for( index_t j =0 ; j< localShape[1] ; j++  )
@@ -18,7 +18,7 @@ void indexDataGenerator::generate( distributedCartesianArray & disData)
                 auto jg = j + localOffset[1];
                 auto kg = k + localOffset[2];
 
-                localData(i,j,k)=ig*ig + jg*jg + kg*kg;
+                localData((signed long int)i,(signed long int)j,(signed long int)k)=ig*ig + jg*jg + kg*kg;
             }   
         }
